PR comments. Summary of the change: - moved UI specific files to `sim-host/ui` - created a new `sim-host/protocol` folder for `socket.js` - handled Chrome-specific CSS inside the `sim-host.css` file
