+++ src/sim-host/protocol/socket.js
