--- src/sim-host/sim-host.css
@@ -1,321 +1,0 @@
-/* Copyright (c) Microsoft Corporation. All rights reserved. */
-
-body {
-    margin: 0 !important;
-    padding: 0 !important;
-    border: 0 !important;
-    font-family: 'Helvetica Neue', 'Roboto', 'Segoe UI', sans-serif;
-    background: rgb(238, 238, 238);
-    user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    -webkit-user-select: none;
-    cursor: default;
-}
-
-html {
-    height: 100%;
-}
-
-h1 {
-    font-family: 'Helvetica Neue', 'Roboto', 'Segoe UI', sans-serif;
-    font-size: 16pt;
-    font-weight: bold;
-}
-
-a {
-    text-decoration: underline;
-    outline: none;
-    cursor: pointer;
-    font-size: 12px;
-    color: rgb(128, 128, 128);
-}
-
-cordova-panel {
-    width: 320px;
-    padding: 5px 0;
-}
-
-cordova-dialog {
-    width: 377px;
-}
-
-cordova-panel, cordova-dialog {
-    z-index: 50;
-    position: relative;
-    -webkit-column-break-inside: avoid;
-    page-break-inside: avoid;
-    break-inside: avoid;
-    display: block;
-}
-
-body /deep/ .cordova-panel-inner {
-    box-shadow: 2px 2px 5px 1px rgba(0, 0, 0, 0.2);
-    background: #fff;
-    border-radius: 3px;
-    overflow: hidden;
-}
-
-body /deep/ .cordova-header {
-    padding: 5px 9px;
-    position: relative;
-    background-color: #03a9f4;
-    color: #fff;
-    font-size: 16px;
-    height: 24px;
-    line-height: 24px;
-}
-
-body /deep/ .cordova-widget {
-    font-family: 'Helvetica Neue', 'Roboto', 'Segoe UI', sans-serif;
-    font-size: 14px;
-}
-
-/* 12px margin around top level items in the content a panel */
-body /deep/ .cordova-content > ::content > * {
-    margin: 12px;
-    position: relative;
-}
-
-.cordova-main {
-    margin-left: 10px;
-    margin-top: 5px;
-    column-width: 330px;
-    -moz-column-width: 330px;
-    -webkit-column-width: 330px;
-    column-gap: 0;
-    -moz-column-gap: 0;
-    -webkit-column-gap: 0;
-}
-
-body /deep/ label, body /deep/ p, body /deep/ .p {
-    color: #3c8b9e;
-    font-size: 14px;
-}
-
-body /deep/ .cordova-value {
-    color: black;
-    font-size: 14px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    margin-left: 6px;
-}
-
-body /deep/ input {
-    margin: 0 10px;
-    padding: 0;
-}
-
-body /deep/ .cordova-overlay {
-    position: fixed;
-    left: 0;
-    top: 0;
-    z-index: 10000;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(245, 245, 245, 0.9);
-
-    display: -webkit-box;
-    display: -moz-box;
-    display: -ms-flexbox;
-    display: -webkit-flex;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-cordova-combo /deep/ select {
-    min-width: 112px;
-}
-
-cordova-button {
-    min-width: 112px;
-    display: inline-block;
-    position: relative;
-    text-align: center;
-    zoom: 1;
-    overflow: visible;
-}
-
-cordova-button /deep/ button {
-    width: 100%;
-    padding: .4em 1em;
-    cursor: pointer;
-}
-
-body /deep/ select {
-    cursor: pointer;
-}
-
-body /deep/ select,
-body /deep/ input[type^=text], input[type^=number] {
-    padding: 5px;
-}
-
-body /deep/ input,
-body /deep/ textarea,
-body /deep/ keygen,
-body /deep/ select,
-body /deep/ button,
-body /deep/ isindex {
-    font-size: 14px;
-    margin-left: 0;
-    margin-right: 0;
-    resize: none;
-}
-
-body /deep/ textarea {
-    padding: 2px 4px;
-    width: calc(100% - 10px); /* Account for padding and border */
-    overflow-y: auto;
-}
-
-body /deep/ input[type^=text],
-body /deep/ input[type^=number] {
-    width: 100px;
-}
-
-body /deep/ button:active,
-body /deep/ textarea:hover,
-body /deep/ textarea:active,
-body /deep/ input[type^=text]:focus,
-body /deep/ input[type^=number]:focus,
-body /deep/ input[type^=text]:hover,
-body /deep/ input[type^=number]:hover {
-    border: 1px solid #CCCCCC;
-    background: #ffffff;
-    font-weight: normal;
-    color: #212121;
-}
-
-body /deep/ .cordova-state-default {
-    border: 1px solid #d3d3d3;
-    background: #e6e6e6;
-    font-weight: normal;
-    color: #555;
-}
-
-body /deep/ input:focus,
-body /deep/ textarea:focus,
-body /deep/ isindex:focus,
-body /deep/ keygen:focus,
-body /deep/ select:focus {
-    outline: none;
-}
-
-body /deep/ cordova-radio {
-    display: flex;
-    align-items: center;
-}
-
-body /deep/ .cordova-panel-row {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-body /deep/ cordova-group, body /deep/ cordova-item-list {
-    display: block;
-}
-
-body /deep/ .cordova-radio-wrapper.cordova-inline {
-    display: inline-flex;
-}
-
-body /deep/ .cordova-radio-label {
-    padding-left: 0.25em;
-}
-
-body /deep/ cordova-item + cordova-item {
-    display: block;
-    border-top: #d3d3d3 1px solid;
-    margin-top: 10px;
-    padding-top: 10px;
-}
-
-body /deep/ cordova-item:hover:before {
-    content: url(delete.png);
-    position: absolute;
-    margin-left: auto;
-    right: 0;
-    background: white;
-    font-size: 14px;
-    padding-left: 8px;
-    font-weight: bold;
-    color: darkred;
-    cursor: pointer;
-}
-
-body /deep/ .cordova-collapse-icon {
-    cursor: pointer;
-    background: url('collapse.png') no-repeat 50% 50%;
-    width: 22px;
-    height: 22px;
-    position: absolute;
-    margin-top: 1px;
-    margin-bottom: auto;
-    right: 6px;
-    margin-left: auto;
-}
-
-body /deep/ .cordova-collapse-icon.cordova-collapsed {
-    background: url('expand.png') no-repeat 50% 50%;
-}
-
-body /deep/ .cordova-collapse-icon:hover, body /deep/ .cordova-close-icon:hover {
-    background-color: rgba(255, 255, 255, 0.25);
-}
-
-body /deep/ .cordova-close-icon {
-    cursor: pointer;
-    background: url('close.png') no-repeat 50% 50%;
-    width: 22px;
-    height: 22px;
-    position: absolute;
-    margin-top: 1px;
-    margin-bottom: auto;
-    right: 6px;
-    margin-left: auto;
-}
-
-body /deep/ .cordova-content {
-    overflow: hidden;
-    transition: height 0.2s;
-}
-
-/* Styles to enable feedback link in a footer panel */
-
-.feedbackFooter {
-    background: rgba(255, 255, 255, 0.85);
-    height: 30px;
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-    box-shadow: 0 -1px 1px 1px rgba(0, 0, 0, 0.2);
-    z-index: 1000;
-}
-
-.feedbackLink {
-    position: absolute;
-    top: 5px;
-    right: 10px;
-}
-
-.cordova-main {
-    padding-bottom: 40px;
-}
-
-/* Hacks to work around multi-column bug in Chrome 48+. The following section
-   is removed for other browsers when the file is served */
-
-/* BEGIN CHROME ONLY */
-body /deep/ .cordova-panel-inner {
-    overflow: visible;
-    padding-bottom: 1px;
-}
-
-body /deep/ .cordova-content {
-    overflow: visible;
-}
-/* END CHROME ONLY */
