+++ src/sim-host/ui/close.png
