+++ src/sim-host/ui/collapse.png
