+++ src/sim-host/ui/delete.png
