+++ src/sim-host/ui/expand.png
